#418 REFACTOR Generate new github client code and fix code to mathc changes

+++ code/src/githubc2/githubc2_components_pending_deployment.ml
@@ -3,7 +3,7 @@ module Primary = struct
     module Primary = struct
       type t = {
         html_url : string option; [@default None]
-        id : int option; [@default None]
+        id : int64 option; [@default None]
         name : string option; [@default None]
         node_id : string option; [@default None]
         url : string option; [@default None]
